+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Support scales "concieved of ... using a division of the perfect fourth interval"
+
 # 0.0.0
 
 Initial gem release.